setup_merged_usr: create skip flag when merged-usr is disabled on bookworm+
usrmerge since version 26 allows to skip the conversion when
installing by creating a specific flag file in /etc.
Create it when --no-merged-usr is used on bookworm and newer, so that
when usrmerge becomes transitively-essential the conversion doesn't
happen if the option was specified.

This will be desirable on buildds for a while, to mitigate packages
that don't work on unmerged-/usr systems if they are built on a
merged-/usr system. --variant=buildd already implies --no-merged-usr
for this reason.
This is also a desirable option to have available for QA systems like
autopkgtest and piuparts, if they want to verify that newly-built
packages can still install and work on unmerged-/usr systems during
the bullseye -> bookworm upgrade. To get this test environment, they
can use --no-merged-usr explicitly.

(cherry picked from commit 43913260cc6a97e0cd86d7d995065f2397fc38e0)